Reduce evaluation allocation
