+++ Ical.Net.Tests/TestHelpers/OccurrenceTester.cs
@@ -56,7 +56,7 @@ internal static class OccurrenceTester
                 {
                     end = start.LocalDateTime
                         .Plus(d.GetNominalPart())
-                        .InZone(start.Zone, Evaluator.ResolveFrom(start))
+                        .InZoneRelativeTo(start)
                         .Plus(d.GetTimePart());
                 }
                 else if (expectedPeriods[i].EndTime is { } periodEnd)
